requirements.txt
git+https://github.com/WAStudios/WASEngine.git@main#egg=WASEngine
lupa
gdown
GitPython
pyyaml
pygame  # Or tkinter/pyqt5 depending on GUI choice